SIMSBIOHUB-668 Refactored as per feedback, improved error messages
