+++ api/src/services/attachment-service.ts
@@ -1,6 +1,7 @@
 import { QueryResult } from 'pg';
-import { ATTACHMENT_TYPE } from '../constants/attachments';
+import { ATTACHMENT_TYPE, TELEMETRY_CREDENTIAL_ATTACHMENT_TYPE } from '../constants/attachments';
 import { IDBConnection } from '../database/db';
+import { ApiGeneralError } from '../errors/api-error';
 import {
   GetAttachmentsWithSupplementalData,
   IReportAttachmentAuthor,
@@ -12,20 +13,53 @@ import {
   IProjectAttachment,
   IProjectReportAttachment,
   IProjectReportAttachmentAuthor,
-  IResponseTelemetryCredentialAttachment,
   ISurveyAttachment,
   ISurveyReportAttachment,
   ISurveyReportAttachmentAuthor,
   SurveyTelemetryCredentialAttachment
 } from '../repositories/attachment-repository';
+import { TelemetryLotekRepository } from '../repositories/telemetry-repositories/telemetry-lotek-repository';
+import { TelemetryVectronicRepository } from '../repositories/telemetry-repositories/telemetry-vectronic-repository';
+import { TelemetryVendorRepository } from '../repositories/telemetry-repositories/telemetry-vendor-repository';
 import { deleteFileFromS3, generateS3FileKey } from '../utils/file-utils';
 import { IValidationData } from '../utils/media/media-utils';
 import { DBService } from './db-service';
 import { HistoryPublishService } from './history-publish-service';
+import { getTelemetryDeviceKey } from './telemetry-services/telemetry-utils';
 
 export interface IAttachmentType {
   id: number;
   type: 'Report' | 'Other';
+}
+
+/**
+ * Response object for inserting a device key into the tables
+ *
+ * @export
+ * @interface IResponseTelemetryCredentialAttachment
+ * @typedef {IResponseTelemetryCredentialAttachment}
+ */
+export interface IResponseTelemetryCredentialAttachment {
+  key?: string;
+  survey_telemetry_credential_attachment_id?: number;
+  survey_telemetry_vendor_credential_id?: number[];
+  telemetry_credential_lotek_id?: number[];
+  telemetry_credential_vectronic_id?: number[];
+}
+
+/**
+ * Data required to persist a device key
+ *
+ * @export
+ * @interface IDeviceKeyData
+ * @typedef {IDeviceKeyData}
+ */
+export interface IDeviceKeyData {
+  fileName: string;
+  fileSize: number;
+  fileData: IValidationData;
+  surveyId: number;
+  key: string;
 }
 
 /**
@@ -37,11 +71,17 @@ export interface IAttachmentType {
  */
 export class AttachmentService extends DBService {
   attachmentRepository: AttachmentRepository;
+  telemetryVectronicRepository: TelemetryVectronicRepository;
+  telemetryLotekRepository: TelemetryLotekRepository;
+  telemetryVendorRepository: TelemetryVendorRepository;
 
   constructor(connection: IDBConnection) {
     super(connection);
 
     this.attachmentRepository = new AttachmentRepository(connection);
+    this.telemetryVectronicRepository = new TelemetryVectronicRepository(connection);
+    this.telemetryLotekRepository = new TelemetryLotekRepository(connection);
+    this.telemetryVendorRepository = new TelemetryVendorRepository(connection);
   }
 
   /**
@@ -927,28 +967,60 @@ export class AttachmentService extends DBService {
   /**
    * Insert survey telemetry credential attachment record.
    *
-   * @param {string} fileName
-   * @param {number} fileSize
-   * @param {IValidationData} fileData
-   * @param {number} surveyId
-   * @param {string} key
-   * @return {*}  {Promise<IResponseTelemetryCredentialAttachment>}
-   * @memberof AttachmentService
+   * @async
+   * @param {IDeviceKeyData} deviceKeyData
+   * @returns {Promise<IResponseTelemetryCredentialAttachment>}
    */
   async insertSurveyTelemetryCredentialAttachment(
-    fileName: string,
-    fileSize: number,
-    fileData: IValidationData,
-    surveyId: number,
-    key: string
+    deviceKeyData: IDeviceKeyData
   ): Promise<IResponseTelemetryCredentialAttachment> {
-    return this.attachmentRepository.insertSurveyTelemetryCredentialAttachment(
-      fileName,
-      fileSize,
-      fileData,
-      surveyId,
-      key
-    );
+    // Initialize empty response json object
+    const responseJSON: IResponseTelemetryCredentialAttachment = {
+      survey_telemetry_vendor_credential_id: [],
+      telemetry_credential_lotek_id: [],
+      telemetry_credential_vectronic_id: []
+    };
+
+    responseJSON.survey_telemetry_credential_attachment_id =
+      await this.attachmentRepository.insertSurveyTelemetryCredentialAttachment(deviceKeyData);
+
+    const vendor = TELEMETRY_CREDENTIAL_ATTACHMENT_TYPE.CFG === deviceKeyData.fileData.type ? 'Lotek' : 'Vectronic';
+    if (deviceKeyData.fileData.keyData) {
+      for (const keyFile of deviceKeyData.fileData.keyData) {
+        // Iterate through the keys array
+        if (keyFile.keysData) {
+          for (const key of keyFile.keysData) {
+            // Generate SIMS device_key
+            const serial = key.id;
+            const deviceKey = getTelemetryDeviceKey({ vendor, serial });
+
+            // populate device key vendor table
+            responseJSON.survey_telemetry_vendor_credential_id?.push(
+              await this.telemetryVendorRepository.insertTelemetryCredentialAttachmentVendor(
+                deviceKey,
+                responseJSON.survey_telemetry_credential_attachment_id
+              )
+            );
+
+            // the data is for lotek cfg
+            if ('Iridium IMEI' in key) {
+              responseJSON.telemetry_credential_lotek_id?.push(
+                await this.telemetryLotekRepository.insertTelemetryCredentialAttachmentLotek(key)
+              );
+            }
+
+            // the data is for vectronic keyx
+            if ('comID' in key && 'comType' in key && 'collarType' in key) {
+              responseJSON.telemetry_credential_vectronic_id?.push(
+                await this.telemetryVectronicRepository.insertTelemetryCredentialAttachmentVectronic(key)
+              );
+            }
+          }
+        }
+      }
+    }
+
+    return responseJSON;
   }
 
   /**
@@ -987,26 +1059,18 @@ export class AttachmentService extends DBService {
     });
 
     const getResponse = await this.getSurveyTelemetryCredentialAttachmentByFileName(file.originalname, surveyId);
-
-    let attachmentResult: IResponseTelemetryCredentialAttachment;
-
     if (getResponse && getResponse.rowCount) {
-      // Existing attachment with matching name found, update it
-      attachmentResult = await this.updateSurveyTelemetryCredentialAttachment(
-        surveyId,
-        file.originalname,
-        attachmentData
-      );
-    } else {
-      // No matching attachment found, insert new attachment
-      attachmentResult = await this.insertSurveyTelemetryCredentialAttachment(
-        file.originalname,
-        file.size,
-        attachmentData,
-        surveyId,
-        key
-      );
+      // Existing attachment with matching name found, throw error
+      throw new ApiGeneralError('Device key file already exists.');
     }
+
+    const attachmentResult = await this.insertSurveyTelemetryCredentialAttachment({
+      fileName: file.originalname,
+      fileSize: file.size,
+      fileData: attachmentData,
+      surveyId: surveyId,
+      key: key
+    });
 
     return { ...attachmentResult, key };
   }
